Provide zero-cost Optioned<char>
This uses `char::from_u32_unsafe(u32::MAX)` as the none value. This is not a valid `char` value, and will never be a valid Unicode code point. Because of this, it is possible to provide a zero cost `Optioned<char>` by assigning the `Noned` value to this invalid code point.

So long as `Optioned` holds its contract of not giving out the noned value, this can in no way be memory unsafe because the fake char is only used as a sentinel, and never as an actual character.